chore(grafana-ui): delete nested component barrel files
