--- packages/grafana-ui/src/components/Button/index.ts
@@ -1,2 +1,0 @@
-export * from './Button';
-export { ButtonGroup } from './ButtonGroup';
