--- packages/grafana-ui/src/components/FileDropzone/index.ts
@@ -1,11 +1,0 @@
-import { FileDropzone, DropzoneFile, FileDropzoneProps, FileDropzoneDefaultChildren } from './FileDropzone';
-import { FileListItem, FileListItemProps } from './FileListItem';
-
-export {
-  FileDropzone,
-  type FileDropzoneProps,
-  type DropzoneFile,
-  FileListItem,
-  type FileListItemProps,
-  FileDropzoneDefaultChildren,
-};
