--- packages/grafana-ui/src/components/Segment/index.ts
@@ -1,7 +1,0 @@
-export { Segment } from './Segment';
-export { SegmentAsync } from './SegmentAsync';
-export { SegmentSelect } from './SegmentSelect';
-export { SegmentInput } from './SegmentInput';
-export { SegmentSection } from './SegmentSection';
-export type { SegmentProps } from './types';
-export { useExpandableLabel } from './useExpandableLabel';
