--- packages/grafana-ui/src/components/SingleStatShared/index.ts
@@ -1,6 +1,0 @@
-export {
-  type SingleStatBaseOptions,
-  sharedSingleStatMigrationHandler,
-  convertOldAngularValueMapping,
-  sharedSingleStatPanelChangedHandler,
-} from './SingleStatBaseOptions';
