--- packages/grafana-ui/src/components/Tabs/index.ts
@@ -1,3 +1,0 @@
-export { Tab } from './Tab';
-export { TabsBar } from './TabsBar';
-export { TabContent } from './TabContent';
