--- packages/grafana-ui/src/components/Toggletip/index.ts
@@ -1,2 +1,0 @@
-export { Toggletip, type ToggletipProps } from './Toggletip';
-export type { ToggletipContent, ToggletipContentProps } from './types';
