--- packages/grafana-ui/src/components/ToolbarButton/index.ts
@@ -1,2 +1,0 @@
-export * from './ToolbarButton';
-export { ToolbarButtonRow } from './ToolbarButtonRow';
