--- packages/grafana-ui/src/components/Tooltip/index.ts
@@ -1,3 +1,0 @@
-export { Tooltip, type TooltipProps } from './Tooltip';
-export { PopoverController } from './PopoverController';
-export type { TooltipPlacement, PopoverContent, PopoverContentProps } from './types';
