--- packages/grafana-ui/src/components/uPlot/geometries/index.ts
@@ -1,4 +1,0 @@
-import { EventsCanvas } from './EventsCanvas';
-import { Marker } from './Marker';
-import { XYCanvas } from './XYCanvas';
-export { XYCanvas, Marker, EventsCanvas };
