--- packages/grafana-ui/src/components/uPlot/plugins/index.ts
@@ -1,3 +1,0 @@
-export { TooltipPlugin2 } from './TooltipPlugin2';
-export { EventBusPlugin } from './EventBusPlugin';
-export { KeyboardPlugin } from './KeyboardPlugin';
